+++ src/main/java/org/example/managers/IPurchaseManager.java
@@ -2,7 +2,7 @@ package org.example.managers;
 
 public interface IPurchaseManager {
 
-    float sumOfMonth (int year, int month);
-    float[] monthlyAverage (int year);
-    float yearlyAverageByCategory(int year);
+    double sumOfMonth (int year, int month);
+    double[] monthlyAverage (int year);
+    double yearlyAverageByCategory(int year);
 }
